refactor: moving commands
