--- src/main/java/sh/sinux/musicmanager/AppCommand/InvalidCommand.java
@@ -1,8 +1,0 @@
-package sh.sinux.musicmanager.AppCommand;
-
-public class InvalidCommand implements Command{
-    @Override
-    public String execute() {
-        return "Invalid command.";
-    }
-}
