+++ src/main/java/sh/sinux/musicmanager/AppCommand/LibraryCommand/InvalidCommand.java
@@ -1,0 +1,10 @@
+package sh.sinux.musicmanager.AppCommand.LibraryCommand;
+
+import sh.sinux.musicmanager.AppCommand.Command;
+
+public class InvalidCommand implements Command {
+    @Override
+    public String execute() {
+        return "Invalid command.";
+    }
+}
