+++ src/main/java/sh/sinux/musicmanager/AppCommand/LibraryCommand/QuitCommand.java
@@ -1,5 +1,6 @@
-package sh.sinux.musicmanager.AppCommand;
+package sh.sinux.musicmanager.AppCommand.LibraryCommand;
 
+import sh.sinux.musicmanager.AppCommand.Command;
 import sh.sinux.musicmanager.Musicmanager;
 
 public class QuitCommand implements Command {
