+++ src/main/java/sh/sinux/musicmanager/AppCommand/LibraryCommand/SizeCommand.java
@@ -1,5 +1,6 @@
-package sh.sinux.musicmanager.AppCommand;
+package sh.sinux.musicmanager.AppCommand.LibraryCommand;
 
+import sh.sinux.musicmanager.AppCommand.Command;
 import sh.sinux.musicmanager.LibraryStorage.LibraryStorage;
 
 public class SizeCommand implements Command {
